// src/app/pages/Cv/enums/color-names.ts
export enum ColorNames {
  Background = 'background',
  CardBackground = 'cardBackground',
  FocusBackground = 'focusBackground',

  Border = 'border',

  Title = 'title',
  Subtitle = 'subtitle',

  Name = 'name',
  Hint = 'hint',
  Text = 'text',

  LanguageFragment = 'languageFragment',
  Function = 'function'
}